pyproject.toml
[project]
name = "point-cloud-filter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "keyboard>=0.13.5",
    "numpy>=2.2.6",
    "open3d>=0.19.0",
    "opencv-python>=4.12.0.88",
    "pyorbbecsdk2>=2.0.15",
    "pyrealsense2>=2.56.5.9235",
    "pyyaml>=6.0.3",
]
